pyproject.toml
[tool.poetry]
name = "asr-server-py"
version = "0.1.0"
description = ""
authors = ["Kai <kai@thekaiway.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
torch = "^2.1.2"
ffmpeg-python = "^0.2.0"
numpy = "^1.26.2"
transformers = "^4.36.1"
gunicorn = "^21.2.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
